Extracted hbar dedicated logic from TransferDetails to HBarTransferDetails subview.
